`useStreamPath` → `getStreamPath` (no react involved)
